HCAP-1214 | moved test to different directory, modified to run with Jest instead of Cypress
